+++ src/client/ui/module/ui_functions.h
@@ -57,7 +57,7 @@ void UI_Init(void);
 void ui_compat_extend_graphics_menu(void);
 void ui_compat_extend_sound_menu(void);
 void ui_compat_refresh_graphics_resolution(void);
-void ui_compat_extend_console_binding_menu(void);
+void ui_compat_extend_misc_binding_menu(void);
 void ui_compat_extend_advanced_menu(void);
 void ui_compat_brand_main_menu_version(void);
 void ui_compat_remove_single_player_menu_item(void);